resolved all the issues in seedEndPoint
