+++ src/modules/admin/seed/controllers/seed.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Post } from "@nestjs/common";
 import { SeedService } from "../services/seed.service";
 import { ApiTags } from "@nestjs/swagger";
 
-@ApiTags('initial data seeding')
+@ApiTags('seed')
 @Controller('seed')
 export class SeedController{
     constructor(private readonly seedService:SeedService){}
