+++ src/modules/admin/seed/data/roles.data.ts
@@ -3,11 +3,18 @@ import { RoleRepository } from '@modules/authorization/repositories/role.reposit
 import { UserTypeEnum } from '@modules/authorization/enums/usertype.enum';
 
 @Injectable()
-export class RoleData {
+export class SeedRoleData {
   constructor(private readonly roleRepository: RoleRepository) {}
 
   async seedRoles(): Promise<void> {
+    const queryRunner = this.roleRepository.getRepository().manager.connection.createQueryRunner();
+
     try {
+      await queryRunner.startTransaction();
+
+      await queryRunner.query('SET FOREIGN_KEY_CHECKS = 0;');
+      await queryRunner.query('TRUNCATE TABLE roles;');
+      await queryRunner.query('SET FOREIGN_KEY_CHECKS = 1;');
       const roles = [
         {
           id: 21,
@@ -26,10 +33,14 @@ export class RoleData {
       ];
 
       await this.roleRepository.getRepository().save(roles);
+      await queryRunner.commitTransaction();
       console.log('Roles table seeding completed.');
     } catch (error) {
+      await queryRunner.rollbackTransaction();
       console.error('Error during roles table seeding:', error);
       throw error;
-    }
+    } finally {
+        await queryRunner.release();
+      }
   }
 }
